fix proxy serialization
